add utility type to improve code model
