+++ community-modules/core/src/ts/gridApi.ts
@@ -62,13 +62,11 @@ import {
     IsRowSelectable,
     IsServerSideGroup,
     IsServerSideGroupOpenByDefaultParams,
-    NavigateToNextCellParams,
     NavigateToNextHeaderParams,
     PaginationNumberFormatterParams,
     PostProcessPopupParams,
     ProcessRowParams,
     ServerSideStoreParams,
-    TabToNextCellParams,
     TabToNextHeaderParams,
     RowClassParams,
     RowHeightParams
@@ -100,6 +98,7 @@ import { OverlayWrapperComponent } from "./rendering/overlays/overlayWrapperComp
 import { HeaderPosition } from "./headerRendering/common/headerPosition";
 import { NavigationService } from "./gridBodyComp/navigationService";
 import { FrameworkComponentWrapper } from "./components/framework/frameworkComponentWrapper";
+import { NavigateToNextCellParams, TabToNextCellParams } from "./entities/iGridCallbacks";
 
 export interface StartEditingCellParams {
     /** The row index of the row to start editing */
