+++ community-modules/core/src/ts/interfaces/iParams.ts
@@ -1,0 +1,17 @@
+import { ColumnApi } from "../columns/columnApi";
+import { GridApi } from "../gridApi";
+
+/**
+ * Enables types safe create of the given type without the need to set the common grid properties
+ * that will be merged with the object in a centralised location.
+ */
+export type WithoutGridCommon<T extends AgGridCommon> = Omit<T, keyof AgGridCommon>;
+
+export interface AgGridCommon {
+    /** The grid api. */
+    api: GridApi;
+    /** The column api. */
+    columnApi: ColumnApi;
+    /** Application context as set on `gridOptions.context`. */
+    context?: any;
+}
